rearranging flow methods
