+++ src/data/run.py
@@ -30,9 +30,9 @@ if not (flow == 'No'):
 
 # Convert masked/flow videos to npz
 if flow == 'Yes':
-    os.system('python to_npz.py --flow=True --smooth=' + str(smooth))
+    os.system('python to_npz_flow.py')
 elif flow == 'No':
     os.system('python to_npz.py --smooth=' + str(smooth))
 else:
-    os.system('python to_npz.py --flow=True --smooth=' + str(smooth))
+    os.system('python to_npz_flow.py')
     os.system('python to_npz.py --smooth=' + str(smooth))
